Add ripemd160 hash function

+++ rmd160.h
@@ -42,7 +42,7 @@
 /* adapt these, if necessary, 
    for your operating system and compiler */
 typedef    unsigned char        byte;
-typedef    unsigned long        dword;
+typedef    unsigned int        dword;
 
 /* if this line causes a compiler error, 
    adapt the defintion of dword above */
